# kask/docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hKask API",
    "description": "A Minimal Viable Container for Replicants — HTTP API.\n\nhKask is an agent runtime grounded in 12 architectural principles\n(P0–P12) expressed through four composable patterns: Skills Model,\nRegulation Feedback Loop, Agentic AI Mediation, and Agent Creation with\nSovereign Memory. This API exposes all capabilities equally across\nCLI, API, and MCP surfaces under P3 (Equal Surface Exposure).\n\nAll endpoints carry OCAP DelegationToken authentication (P4).\nData access is governed by user sovereignty and affirmative\nconsent (P1, P2).",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.30.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "hKask API server"
    }
  ],
  "paths": {
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat with the Curator or a specified agent.",
        "description": "Accepts an optional `model` field to switch the LLM at request time.\nWhen omitted, the server default (a fast classifier model) is used. The response\nechoes the `model` used, confirming which LLM generated the output.\n\nUse `GET /api/models` or `GET /api/models/search?q=...` to discover\navailable model identifiers.",
        "operationId": "chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Chat response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Stream chat inference output as Server-Sent Events.",
        "description": "Accepts the same `ChatRequest` as `/api/chat` but streams the inference\noutput as SSE `data:` events. Each event carries a JSON object with\n`text_delta`, `model`, and optionally `finish_reason`. The final event\nhas `finish_reason: \"stop\"` and includes `usage` stats.\n\nThis is a surface-layer endpoint — it bypasses ChatService's full\npipeline (memory recall, episodic storage) and calls the inference port\ndirectly for streaming. Use `/api/chat` for the full pipeline with\nmemory integration.",
        "operationId": "chat_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE stream of chat chunks",
            "content": {
              "text/event-stream": {}
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List all available models from all configured providers.",
        "description": "Queries DeepInfra, fal.ai, Together AI, OpenRouter, and KiloCode and returns metadata for each\navailable model with provider prefix applied. Returns an empty list if\nno providers are reachable (graceful degradation).",
        "operationId": "list_models",
        "responses": {
          "200": {
            "description": "List of available models",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            }
          },
          "503": {
            "description": "No providers reachable (returns empty list)"
          }
        }
      }
    },
...
    "/api/v1/a2a/agents": {
      "get": {
        "tags": [
          "a2a"
        ],
        "summary": "List all registered A2A agents",
        "operationId": "a2a_list_agents",
        "responses": {
          "200": {
            "description": "List of registered agents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/a2a/agents/{agent_id}": {
      "delete": {
        "tags": [
          "a2a"
        ],
        "summary": "Unregister an A2A agent",
        "operationId": "a2a_unregister_agent",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "description": "A2A agent ID to unregister",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent unregistered"
          },
          "400": {
            "description": "Invalid agent ID"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Agent not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "A2AAgentResponse": {
        "type": "object",
        "description": "A2A agent response — registered agent in the ACP capability delegation system (P4 OCAP).\n\n`agent_type` is \"Bot\" or \"Replicant\" (P10).\n`capabilities` are the granted capability verbs this agent holds.\n`active` indicates whether the agent is currently allowed to exercise its capabilities.",
        "required": [
          "webid",
          "agent_type",
          "capabilities",
          "registered_at",
          "active"
        ],
        "properties": {
          "active": {
            "type": "boolean"
          },
          "agent_type": {
            "type": "string"
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "registered_at": {
            "type": "integer",
            "format": "int64"
          },
          "webid": {
            "type": "string"
          }
        }
      },
...
      "AgentListResponse": {
        "type": "object",
        "description": "Agent list response — all A2A-registered agents.",
        "required": [
          "agents"
        ],
        "properties": {
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/A2AAgentResponse"
            }
          }
        }
      },
...
      "ChatRequest": {
        "type": "object",
        "description": "Chat request sent to the Curator or a specified agent.\n\nThe `model` field allows switching the LLM at request time. When omitted,\nthe server default (a fast classifier model) is used. Use `GET /api/models` to discover\navailable models, and `GET /api/models/search?q=...` for fuzzy matching.",
        "required": [
          "input"
        ],
        "properties": {
          "input": {
            "type": "string",
            "description": "User input message"
          },
          "model": {
            "type": [
              "string",
              "null"
            ],
            "description": "Model override for inference (e.g., \"DI/google/gemma-4-9b-it\"). If unset, uses the server default."
          },
          "template_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional template ID to contextualize the prompt"
          }
        }
      },
      "ChatResponse": {
        "type": "object",
        "description": "Chat response from the Curator or agent.\n\nThe `model` field echoes which LLM was used, confirming model switching.",
        "required": [
          "output",
          "template_id",
          "model"
        ],
        "properties": {
          "model": {
            "type": "string",
            "description": "Model identifier used for inference"
          },
          "output": {
            "type": "string",
            "description": "Generated response text"
          },
          "template_id": {
            "type": "string",
            "description": "Template ID that was applied (or \"auto-select\")"
          }
        }
      },
...
      "ModelEntry": {
        "type": "object",
        "description": "A model available through the inference router.\n\nIncludes metadata from the provider's model listing endpoint:\nmodel family, parameter count, quantization level, and disk size.",
        "required": [
          "name"
        ],
        "properties": {
          "family": {
            "type": [
              "string",
              "null"
            ],
            "description": "Model family (e.g., \"llama\", \"qwen2\")"
          },
          "name": {
            "type": "string",
            "description": "Model identifier (e.g., \"qwen3:8b\", \"llama3.1:70b\")"
          },
          "parameter_size": {
            "type": [
              "string",
              "null"
            ],
            "description": "Parameter count (e.g., \"8B\", \"70B\")"
          },
          "quantization_level": {
            "type": [
              "string",
              "null"
            ],
            "description": "Quantization level (e.g., \"Q4_0\", \"Q5_K_M\")"
          },
          "size_gb": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Model size in gigabytes"
          }
        }
      },
      "ModelListResponse": {
        "type": "object",
        "description": "Response containing available models.\n\nReturned by `GET /api/models` and `GET /api/models/search`.\nModel names from this response can be used as the `model` field\nin `POST /api/chat` requests.",
        "required": [
          "models",
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "description": "Total number of models in the list",
            "minimum": 0
          },
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelEntry"
            },
            "description": "List of available models"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "DelegationToken",
        "description": "DelegationToken — an OCAP capability token carrying the authenticated WebID and scoped permissions (P4).\n\nObtain via the REPL onboarding flow (`kask secret`) or agent registration (A2A)."
      }
    }
  },
  "security": [
    {
      "bearer_token": []
    }
  ],
...
}